+++ tests/utils/test_cache.py
@@ -3,7 +3,7 @@ from typing import Dict
 
 from BALSAMIC.constants.cache import ContainerVersion
 
-from BALSAMIC.commands.init.utils import get_containers
+from BALSAMIC.utils.cache import get_containers
 
 
 def test_get_containers(develop_containers: Dict[str, str]):