+++ CHANGELOG.rst
@@ -7,9 +7,12 @@ Added:
 * `CADD` SNV annotation https://github.com/Clinical-Genomics/BALSAMIC/pull/1150
 * Samtools stats, flagstat, idxstat to WGS workflow
 * Functionality for dynamically assigning fastq-info to sample-dictionary from input fastq-dir
-* Annotate SNVs with cancer germline SNV observations https://github.com/Clinical-Genomics/BALSAMIC/pull/1178
-* Annotate SNVs with somatic SNV observations https://github.com/Clinical-Genomics/BALSAMIC/pull/1187
+* Annotate SNVs with cancer germline SNV observations from loqusDB https://github.com/Clinical-Genomics/BALSAMIC/pull/1178
+* Annotate SNVs with somatic SNV observations from loqusDB https://github.com/Clinical-Genomics/BALSAMIC/pull/1187
 * Tests for Annotation with Cancer germline, somatic and clinical observations, and swegen frequencies https://github/Clinical-Genomics/BALSAMIC/pull/1190
+* Annotate SVs with somatic SV observations from loqusDB https://github.com/Clinical-Genomics/BALSAMIC/pull/1194
+* Support singularity bind paths with different destination directories https://github/Clinical-Genomics/BALSAMIC/pull/1211
+* Added `--rerun-trigger mtime` option to Snakemake command https://github.com/Clinical-Genomics/BALSAMIC/pull/1217
 
 Changed:
 ^^^^^^^^
@@ -18,10 +21,13 @@ Changed:
 * Standardised alignment workflows to WGS-workflow
 * Implemented parallel trimming and alignment in all workflows per lane
 * All bam-QC tools take the final dedup.realign bamfile as input
+* Validation of pydantic models done both during config and run
 * Fix pydantic version (<2.0) https://github.com/Clinical-Genomics/BALSAMIC/pull/1191
 * Refactor constants https://github.com/Clinical-Genomics/BALSAMIC/pull/1174
 * Move models to their own folder https://github.com/Clinical-Genomics/BALSAMIC/pull/1176
 * Balsamic init workflow refactoring https://github.com/Clinical-Genomics/BALSAMIC/pull/1188
+* Updated cryptography version to 41.0.2 https://github.com/Clinical-Genomics/BALSAMIC/pull/1205
+* Refactor snakemake executable command generation https://github/Clinical-Genomics/BALSAMIC/pull/1211
 
 Fixed:
 ^^^^^^
@@ -29,6 +35,7 @@ Fixed:
 * Link external fastqs to case folder & create case directory https://github.com/Clinical-Genomics/BALSAMIC/pull/1195
 * vcf2cytosure container missing constants https://github.com/Clinical-Genomics/BALSAMIC/pull/1198
 * Bash commands in vep_somatic_clinical_snv https://github.com/Clinical-Genomics/BALSAMIC/pull/1200
+* Fix SVDB annotation intermediate rule https://github.com/Clinical-Genomics/BALSAMIC/pull/1218
 
 Removed:
 ^^^^^^^
@@ -41,7 +48,6 @@ Fixed:
 ^^^^^^
 * Missing `Number` in VCF header for SVs  https://github.com/Clinical-Genomics/BALSAMIC/pull/1203
 
-=======
 Changed:
 ^^^^^^^^
 * Fix cyvcf2 to version 0.30.22 https://github.com/Clinical-Genomics/BALSAMIC/pull/1206